+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-27 03:11:34
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 8abfb846e14d5fdb4ff3b1aa2b797758ebb89544
+
+## [ForwardWidgets-2kuai] - 2025-10-27 03:11:34
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: ded56cb429e7ac925d7a101b736a9c91d1942717
+
+## [ForwardWidgets-pack1r] - 2025-10-27 03:11:34
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 528e140ba675a201660c848391eb27e95f7747c1
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-26 03:02:37
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main